test(rss): fold tests into sibling tests/ submodule per topic-split convention
services/rss/mod.rs sat at 2351 LoC, well past the ~1500 threshold
where tests/CLAUDE.md says inline tests should move to a sibling
tests/ directory. The two pre-existing test files
(evaluate_candidate_tests.rs and parse_release_tests.rs) were
already siblings to mod.rs but outside the convention's directory
shape, and a third inline `mod source_tests { ... }` block lived
in mod.rs itself.

Moves:
- evaluate_candidate_tests.rs → tests/evaluate_candidate.rs
- parse_release_tests.rs → tests/parse_release.rs
- inline source_tests block → tests/source.rs

Wires up via a new tests/mod.rs that declares the three submodules,
plus the standard `#[cfg(test)] mod tests;` declaration in
rss/mod.rs (replacing the three previous mod declarations).

Drops the redundant `_tests` suffix from filenames since the
parent directory is now `tests/` — matches the existing
post_processing/tests/ convention (file_ops.rs, filenames.rs,
lock.rs).

`super::*` imports inside the moved files become `super::super::*`
since they're now one level deeper in the module tree.

Verified clean: cargo check + cargo nextest run (2142 passed,
23 skipped) + cargo fmt + cargo clippy --all-targets -D warnings.
