+++ src/services/rss/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Topic-split test submodule for `services::rss` (per the convention
+//! in `tests/CLAUDE.md` — sibling `tests/` directory once the parent
+//! module exceeds ~1500 LoC).
+
+mod evaluate_candidate;
+mod parse_release;
+mod source;
